Update custom_css.txt
I'm not sure what's going on but the google search box styles aren't always been set when the search box is moved into position on my Samsung phone. I have yet to see it happen when the styles are in the page as opposed to them being in a separate style sheet. I'm wondering if it is to do with when the javascript code kicks in for changing the image or moving the search box. I don't think I've seen it on any other device yet.