Update footer_content_not_minimised
Changed how the directory works. Using a closure so I can iterate over a list of the panel links to add event listeners to them. It also meant changing the rest of the panel code to deal with the element being passed in instead of the id as a string. I also changed how the panel links are named, they now have to have a hyphen to separate the anchor part from the panel name. I am also passing in the array of panel names to make the code more generic so that other panels can potentially use this code (not that there are any other panels in the site currently).