Initial project build setup
Setup project to use gulp and nom package manager from NodeJS.  This will make it easier to manage content in components of functionality with spacing and appropriate comments but have a build process that automates the building of files to be added to SwimTopia website advanced site settings.  Specifically the build will generate a dist (distribution) directory which currently have files to be copied into the “Additional Header Content (Advanced)” section and the “Footer Content” sections.  Header content includes all custom CSS and Javascript concatenated and minified.